requirements.txt
Pillow==8.2.0
pytesseract==0.3.7
